requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.0
django-tastypie==0.14.0
flup6==1.1.1
idna==2.6
mysqlclient==1.3.12
oauthlib==2.0.6
PyJWT==1.5.3
pyparsing==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
scss==0.8.73
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
urllib3==1.22
